pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ia_modules"
version = "0.0.4"
description = "Production-ready AI agent framework with enterprise-grade reliability and observability"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "IA Modules Contributors"}
]
keywords = [
    "ai", "agents", "pipeline", "workflow", "orchestration",
    "reliability", "observability", "earf", "llm"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "aiosqlite>=0.19.0",
    "pydantic>=2.0.0",
    "fastapi>=0.104.0",
    "bcrypt>=4.0.0",
    "nexusql>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "python-dotenv>=1.0.0",
]
redis = [
    "redis>=5.0.0",
]
postgres = [
    "psycopg2-binary>=2.9.0",
]
mysql = [
    "pymysql>=1.1.0",
]
mssql = [
    "pyodbc>=5.0.0",
]
sql = [
    "psycopg2-binary>=2.9.0",  # PostgreSQL
    "pymysql>=1.1.0",  # MySQL
    "pyodbc>=5.0.0",  # MSSQL
]
observability = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-exporter-prometheus>=0.41b0",
    "opentelemetry-instrumentation>=0.41b0",
    "prometheus-client>=0.19.0",
]
cli = [
    "graphviz>=0.20.0",
    "click>=8.0.0",
]
profiling = [
    "psutil>=5.9.0",
]
# Advanced AI features
guardrails = [
    "openai>=1.0.0",
]
rag-advanced = [
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]
agents-advanced = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]
ai-advanced = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]
all = [
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.0",
    "pymysql>=1.1.0",
    "pyodbc>=5.0.0",
    "graphviz>=0.20.0",
    "click>=8.0.0",
    "psutil>=5.9.0",
    "pytest-cov>=4.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-exporter-prometheus>=0.41b0",
    "opentelemetry-instrumentation>=0.41b0",
    "prometheus-client>=0.19.0",
    # Advanced AI features
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]

[project.urls]
Homepage = "https://github.com/Fiberwise-AI/ia_modules"
Documentation = "https://ia-modules.readthedocs.io"
Repository = "https://github.com/Fiberwise-AI/ia_modules"
Issues = "https://github.com/Fiberwise-AI/ia_modules/issues"

[project.scripts]
ia-modules = "ia_modules.cli.main:main"

[tool.setuptools]
packages = ["ia_modules"]
package-dir = {"ia_modules" = "."}

[tool.setuptools.package-data]
ia_modules = ["*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
filterwarnings = [
    "ignore:The default datetime adapter is deprecated:DeprecationWarning",
]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
    "redis: marks tests that require Redis server",
    "postgres: marks tests that require PostgreSQL server",
    "mysql: marks tests that require MySQL server",
    "mssql: marks tests that require MSSQL server",
    "observability: marks tests that require observability stack (Prometheus, Grafana, etc)",
]

[tool.coverage.run]
source = [
    "ia_modules",
]
# Coverage includes all submodules:
# agents, auth, benchmarking, checkpoint, cli, database,
# memory, pipeline, plugins, rag, reliability, scheduler,
# telemetry, tools, validation
omit = [
    # Test files and directories
    "*/tests/*",
    "*/test_*.py",
    "*/conftest.py",
    "tests/*",

    # Build and cache directories
    "*/__pycache__/*",
    "*.egg-info/*",

    # Excluded apps and utilities
    "showcase_app/*",
    "archive/*",
    "docs/*",
    "release_notes/*",

    # Example and manual test scripts
    "tests/llm_config_example.py",
    "tests/manual_database_test.py",
    "tests/pipeline_runner.py",
    "tests/pipeline_runner_with_llm.py",
]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
precision = 2
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"